Added project
This project is still a work in progress and misses several parts: The chapters introduction, tech specs, how it works, process, software and hardware need (more) text. There are also several images and a video missing for introduction, design, parts of the process, software and hardware and we also still need a name. The missing parts will be added as soon as the needed information is provided.